requirements.txt
# Backend Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database & ORM
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.0

# Frontend
streamlit==1.37.0
plotly==5.17.0
pandas==2.1.3

# Database Drivers
psycopg2-binary==2.9.9

# Background Tasks & Caching - FIXED VERSIONS
redis==4.6.0
celery[redis]==5.3.4

# API Integration
httpx==0.25.2
requests==2.32.4

# Data Processing
openpyxl==3.1.2
reportlab==4.0.7
pillow==10.3.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.18

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development Tools
python-dotenv==1.0.0

# System Monitoring (for Pi)
psutil==5.9.6

# Task Scheduling
schedule==1.2.0

# Data Science (for analytics)
scikit-learn==1.5.0
matplotlib==3.8.2

# Currency/Exchange Rates
forex-python==1.8

# Memory profiling (development)
memory-profiler==0.61.0

#Performance
watchdog==6.0.0